requirements.txt
requests==2.32.3
urllib3==2.2.3
charset-normalizer==3.4.1
python-dotenv==1.0.1
pandas==2.2.3
tabulate==0.9.0
fastapi==0.115.8
uvicorn==0.34.0
python-telegram-bot==21.10